requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.97.0
pandas==2.3.1
pgvector==0.4.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
SQLAlchemy==2.0.41
starlette==0.47.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
